replace headers for modules in boot and kernel
