+++ src/os/boot/idt/pic/programmable_interrupt_controller.cxx
@@ -15,14 +15,38 @@
  * along with HephaestOS.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "idt/pic/programmable_interrupt_controller.h"
+module;
 
-#include "io.h"
+#include <cstdint>
+
+import os.hal.io;
+
+export module os.boot.interrupts.pic;
 
 namespace boot {
 
+    constexpr uint8_t masterPicAddress = 0x20;// IO Address of Master PIC
+    constexpr uint8_t slavePicAddress = 0xA0;// IO Address of Slave PIC
+    constexpr uint8_t masterPicDataAddress = 0x21;// IO Data Address of Master PIC
+    constexpr uint8_t slavePicDataAddress = 0xA1;// IO Data Address of Slave PIC
+
+
+    constexpr uint8_t ICW1_ICW4 = 0x01; /* ICW4 (not) needed */
+    constexpr uint8_t ICW1_SINGLE = 0x02; /* Single (cascade) mode */
+    constexpr uint8_t ICW1_INTERVAL4 = 0x04; /* Call address interval 4 (8) */
+    constexpr uint8_t ICW1_LEVEL = 0x08; /* Level triggered (edge) mode */
+    constexpr uint8_t initializeCommand = 0x10;// Initialize the PIC
+
+    constexpr uint8_t ICW4_8086 = 0x01; /* 8086/88 (MCS-80/85) mode */
+    constexpr uint8_t ICW4_AUTO = 0x02; /* Auto (normal) EOI */
+    constexpr uint8_t ICW4_BUF_SLAVE = 0x08; /* Buffered mode/slave */
+    constexpr uint8_t ICW4_BUF_MASTER = 0x0C; /* Buffered mode/master */
+    constexpr uint8_t ICW4_SFNM = 0x10; /* Special fully nested (not) */
+
+    constexpr uint8_t picEoiFlag = 0x20;
+
     // Maps the overlapping interrupts to after the exceptions.
-    void remapProgrammableInterruptController(uint8_t masterOffset, uint8_t slaveOffset) {
+    export void remapProgrammableInterruptController(uint8_t masterOffset, uint8_t slaveOffset) {
         // initialize pic todo: Comment second command
         hal::writeToPort(masterPicAddress, initializeCommand | ICW1_ICW4);
         hal::waitForIO();
@@ -59,7 +83,7 @@ namespace boot {
      *
      * @param interruptCode
      */
-    void sendEoiFlag(uint32_t interruptCode) {
+    export void sendEoiFlag(uint32_t interruptCode) {
         if (interruptCode >= 8) {
             hal::writeToPort(slavePicAddress, picEoiFlag);
         }
