+++ src/os/boot/paging/model/page_directory_entry.cxx
@@ -15,14 +15,15 @@
 // along with HephaestOS.  If not, see <https://www.gnu.org/licenses/>.
 //
 
-#ifndef HEPHAEST_OS_KERNEL_BOOT_PAGING_PAGE_DIRECTORY_ENTRY_H
-#define HEPHAEST_OS_KERNEL_BOOT_PAGING_PAGE_DIRECTORY_ENTRY_H
+module;
 
 #include <cstdint>
 
-#include "page_table_entry.h"
+export import os.boot.paging.table;
 
-namespace boot {
+export module os.boot.paging.directory;
+
+export namespace boot {
 
     // Defines the access flags of a given Page Directory Entry.
     struct [[gnu::packed]] PageDirectoryAccess {
@@ -47,4 +48,3 @@ namespace boot {
 
 }// namespace boot
 
-#endif// HEPHAEST_OS_KERNEL_BOOT_PAGING_PAGE_DIRECTORY_ENTRY_H
