--- src/os/boot/paging/model/paging_constants.h
@@ -1,33 +1,0 @@
-// Copyright (C) 2023 David Price - All Rights Reserved
-// This file is part of HephaestOS.
-//
-// HephaestOS is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// HephaestOS is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with HephaestOS.  If not, see <https://www.gnu.org/licenses/>.
-//
-
-#ifndef HEPHAEST_OS_KERNEL_BOOT_PAGING_MODEL_PAGING_CONSTANTS_H
-#define HEPHAEST_OS_KERNEL_BOOT_PAGING_MODEL_PAGING_CONSTANTS_H
-
-#include "cstdint"
-
-namespace boot {
-
-    // Page directory
-    static constexpr std::size_t PAGE_DIRECTORY_SIZE = 1024;
-
-    static constexpr std::size_t PAGE_TABLE_SIZE = 1024;
-
-    static constexpr std::size_t PAGE_SIZE = 0x1000;
-}// namespace boot
-
-#endif// HEPHAEST_OS_KERNEL_BOOT_PAGING_MODEL_PAGING_CONSTANTS_H
